requirements.txt
django
django-cors-headers
django-compat
djangorestframework
django-apscheduler
channels
mysqlclient
argon2-cffi
ultralytics==8.0.184
opencv-python
daphne
waitress
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.0+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
torchvision==0.15.0+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
torchaudio==2.0.0+cu118